remove AuthUtil
